+++ package.json
@@ -9,9 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@nextui-org/react": "^2.4.8",
+    "@nextui-org/react": "^2.6.10",
     "firebase": "^11.0.2",
-    "framer-motion": "^11.11.17",
+    "framer-motion": "^11.15.0",
     "next": "14.2.15",
     "next-themes": "^0.4.3",
     "react": "^18",
@@ -24,7 +24,7 @@
     "eslint": "^8",
     "eslint-config-next": "14.2.15",
     "postcss": "^8",
-    "tailwindcss": "^3.4.1",
+    "tailwindcss": "^3.4.17",
     "typescript": "^5"
   }
 }
